requirements.txt
ebooklib
Pillow
requests
beautifulsoup4
lxml 